requirements.txt
Django==4.2.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Pillow==9.2.0
psycopg2-binary==2.9.9
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2022.4
requests==2.28.1